Move AdapterContainer to lewis.core.adapters
